pythonServer/pyproject.toml
[project]
name = "pythonserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "langchain>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.1",
    "tensorflow>=2.20.0",
]
